requirements.txt
click==8.1.3
greenlet==2.0.2
importlib-metadata==6.3.0
inflect==6.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
pydantic==1.10.7
PyMySQL==1.0.3
PyYAML==6.0
SQLAlchemy==1.4.47
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
